Split up funcitons into separate files
